requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.24
filelock==3.29.0
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.4.0
greenlet==3.4.0
groq==0.37.1
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.23.4
idna==3.12
iniconfig==2.3.0
intel-openmp==2021.4.0
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
kiwisolver==1.5.0
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-groq==0.1.10
langchain-huggingface==0.0.3
langchain-protocol==0.0.15
langchain-text-splitters==0.2.4
langgraph-checkpoint==4.0.3
langgraph-sdk==0.3.13
langsmith==0.1.147
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.9.0
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.6.1
numpy==1.26.4
openai==1.30.1
orjson==3.11.8
ormsgpack==1.12.2
packaging==24.2
pandas==2.2.2
pillow==12.2.0
pluggy==1.6.0
propcache==0.4.1
pydantic==2.7.1
pydantic-settings==2.14.0
pydantic_core==2.18.2
Pygments==2.20.0
pyparsing==3.3.2
pytest==8.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rich==15.0.0
rich-toolkit==0.19.7
safetensors==0.7.0
scikit-learn==1.5.0
scipy==1.17.1
seaborn==0.13.2
sentence-transformers==5.4.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==0.37.2
sympy==1.14.0
tbb==2021.13.1
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.67.3
transformers==4.41.2
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
ujson==5.12.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.30.0
watchfiles==1.1.1
websockets==16.0
xxhash==3.7.0
yarl==1.23.0
zstandard==0.25.0
